rename fileName
* <tpm_core_node>: rename files under ros_interface

* <tpm_controller>: remove warning

+++ tpm_controller/src/FJStrategy_mclcPVT.cpp
@@ -111,7 +111,7 @@ std::vector<double> FJStrategy_mclcPVT::GetMJointStates()
     _mJointStates.clear();
     for (size_t i=0;i<6;i++)
     {
-        double pos;
+        FLT pos;
         mclc_axis_get_trgPosCmd(i, &pos);
         _mJointStates.push_back(pos);
     }
